apps/ml-service/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
pydantic>=2.10.0,<3.0.0
opencv-python>=4.10.0.84,<5.0.0.0
numpy>=2.1.0,<3.0.0
pillow>=10.4.0,<12.0.0
python-dotenv>=1.0.1,<2.0.0
